+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - New: `sized(decoder, {min?, max?, size?})` utility for adding length validation to
   strings, arrays, and sets.
-- New: `min(decoder, min)`, `max(decoder, max)`, and `between(decoder, lo, hi)` utilities
+- New: `min(decoder, min)`, `max(decoder, max)`, and `between(decoder, min, max)` utilities
   for adding inclusive numeric range validation.
 - New: `urlString` decoder, which is like `url`, but returns a string instead of a URL
   instance.